Add radios and buttons.

+++ front/App.vue
@@ -4,10 +4,13 @@
       <myHeader></myHeader>
     </el-header>
     <el-main style="padding: 50px">
+      <el-radio v-model="radio" label="25" :disabled="radio25" @change="xxx">25分钟</el-radio>
+      <el-radio v-model="radio" label="40" :disabled="radio40" @change="xxx">40分钟</el-radio>
       <div>{{this.timeRemain}}</div>
       <div>{{this.remainTimestamp}}</div>
-      <el-button plain @click="start">开始</el-button>
-      <el-button plain @click="stop">停止</el-button>
+
+      <el-button plain @click="start" :disabled="btnStart">开始</el-button>
+      <el-button plain @click="stop" :disabled="btnStop">停止</el-button>
     </el-main>
     <el-footer>Footer</el-footer>
   </el-container>
@@ -22,10 +25,14 @@ export default {
   name: "app",
   data() {
     return {
-      timeRemain: "00:00",
+      radio: "25",
+      radio25: false,
+      radio40: false,
+      btnStart: false,
+      btnStop: false,
+      timeRemain: "25:00",
       remainTimestamp: 0,
-      duration: 0,
-      worker: new Worker()
+      duration: 25
     };
   },
   beforeDestroy() {
@@ -34,17 +41,19 @@ export default {
   components: {
     myHeader
   },
-  created() {
-    this.$bus.$on("select25", data => {
-      this.duration = data.value;
-    });
-    this.$bus.$on("select40", data => {
-      this.duration = data.value;
-    });
-  },
+  created() {},
   mounted() {},
   methods: {
+    xxx(label) {
+      if (label === "25") {
+        this.duration = 25;
+      } else {
+        this.duration = 40;
+      }
+      this.timeRemain = this.duration + ":00";
+    },
     setTimer(val) {
+      this.worker = new Worker();
       this.worker.postMessage({ value: val * 60 * 1000 });
       const that = this;
       this.worker.onmessage = function(e) {
@@ -55,10 +64,25 @@ export default {
       };
     },
     start() {
+      this.btnStart = true;
+      if (this.duration === 25) this.radio40 = true;
+      else this.radio25 = true;
       this.setTimer(this.duration);
     },
     stop() {
-      this.worker.terminate();
+      this.$confirm("剩余时间将被重置，并且也不会得到番茄，确定停止？", "", {
+        confirmButtonText: "确定",
+        cancelButtonText: "取消",
+        type: "warning"
+      })
+        .then(() => {
+          this.worker.terminate();
+          this.timeRemain = this.duration + ":00";
+          this.radio40 = false;
+          this.radio25 = false;
+          this.btnStart = false;
+        })
+        .catch(() => {});
     }
   }
 };
